Implement user management

+++ templates/easy-rsa/gen-vpn.sh
@@ -2,13 +2,16 @@
 
 ## Generate VPN config file from template
 
-source vars
+source ./vars
 
 [ -z "$1" ] && echo "Usage: $0 [key_name]" && exit 1
-[ -z "$2" ] && echo "Usage: $0 [key_name]" && exit 1
+
+if [ -f "generated/$1.ovpn" ] && [ "$2" != 'yes' ]; then
+    echo "User profile exist. Skiping ..."
+    exit 0
+fi
 
 yes | ./build-key $1
-
 
 cp vpn/template.ovpn generated/${1}.ovpn
 echo "<ca>"  >> generated/${1}.ovpn
@@ -29,5 +32,5 @@ echo "<key>" >> generated/${1}.ovpn
 cat $KEY_DIR/$1.key >> generated/${1}.ovpn
 echo "</key>" >> generated/${1}.ovpn
 
-useradd $1
-echo "$1:$2" | chpasswd
+#useradd $1
+#echo "$1:$2" | chpasswd
